+++ docs/CHANGE_LOG.md
@@ -5,6 +5,9 @@
 #### **Binding Options:**
 - Added a new binding option called "mapTogglesEnabled", which states if the map toggles are enabled (defaults to true).
 
+#### **Public Functions:**
+- Added new public function "setYear()", which will set the year being displayed for a heat map DOM element.
+
 #### **General Improvements:**
 - The heat map layout is now fully responsive with mobile and tablet views.
 - When exporting the map details, the type selected is now included in the filename (if types are available).